Add Github CI (#20)
This PR adds GitHub CI to the project that checks that the package
builds. It turns out that there was a typo in the import statements that
was causing build errors (which I dicovered when manually building
locally). In order for CI to pass, the typo is fixed.

In addition, this PR also bumps the package version to v2.1.1 to make a
new bugfix release including the import path update.